4.21 complete, deletion only works for the user that created the blog

+++ part4/blog/controllers/blogs.js
@@ -8,7 +8,7 @@ const User = require('../models/user')
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
     .find({})
-    .populate('user')
+    .populate('user', {blogs: 0})
   response.json(blogs)
 })
 
@@ -50,6 +50,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({error: 'token invalid'})
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if(!blog) {
+    return response.status(404).json({error: 'blog not found'})
+  }
+
+  if(blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(403).json({error: 'only blog creator may delete blog'})
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
